backend/requirements.txt
langchain
llama-cpp-python
faiss-cpu
sentence-transformers
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
